fixed issues for playlists details request
