+++ src/app/controllers/gets.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 /* eslint-disable import/prefer-default-export */
-import { fetchPlaylists } from '../helpers';
+import { fetchPlaylists, clientLogin, updateENV } from '../helpers';
 
 const getPlaylists = async (req, res) => {
   const playlists = await fetchPlaylists()
@@ -12,4 +12,16 @@ const getPlaylists = async (req, res) => {
   res.send(playlists);
 };
 
-export { getPlaylists };
+const getToken = async (req, res) => {
+  const token = await clientLogin()
+    .then((data) => {
+      console.log('processing data..');
+      console.log(data);
+      return data;
+    })
+    .catch((err) => console.log(err));
+  updateENV('TOKEN', token);
+  res.send(token);
+};
+
+export { getPlaylists, getToken };
