+++ src/app/helpers.js
@@ -20,6 +20,7 @@ const authString = Buffer.from(
 
 console.log(authString);
 
+// generate axios json for POST, GET requests
 const apiReqJson = ({
   method = 'GET',
   data = {},
@@ -49,6 +50,7 @@ const apiReqJson = ({
   return finalJson;
 };
 
+// get login token
 const clientLogin = async () => {
   const data = await axios(
     tokenURI,
@@ -67,6 +69,7 @@ const clientLogin = async () => {
   return data;
 };
 
+// fetching user playlists
 const fetchPlaylists = async () => {
   console.log('fetching playlists...');
   console.log(process.env.TOKEN);
@@ -89,6 +92,10 @@ const fetchPlaylists = async () => {
   return data;
 };
 
+/*
+Normalizing port value coming from environment
+@param val: value of incoming port
+*/
 const normalizePort = (val) => {
   const port = parseInt(val, 10);
   if (Number.isNaN) {
@@ -100,8 +107,13 @@ const normalizePort = (val) => {
   return false;
 };
 
+/*
+update environment variable value
+@param key: environment variable name
+@param value: environment variable value
+*/
 const updateENV = (key, value) => {
-  console.log(`updating ${key}..\n`);
+  console.log(`\n updating ${key}..with value ${value}\n`);
   const sourcePath = path.resolve('./src', '../.env');
   const source = fs.readFileSync(sourcePath, { encoding: 'utf8', flag: 'r' });
   console.log('source', source);
