--- supervisord.conf
@@ -1,5 +1,0 @@
-[program:node]
-directory=/srv/www
-command=npm run demo
-autostart=true
-autorestart=true
